requirements.txt
# Core
python-dotenv>=1.0.0

# Vector DB & embeddings
pinecone-client>=3.0.0
sentence-transformers>=2.0.0

# LangChain & OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Streamlit app (langchainbot, chatbot)
streamlit>=1.28.0

# Optional: for RAGDemo.ipynb
# langchain-chroma
# langchain-text-splitters
# beautifulsoup4

# Optional: for similarity_search.py
# transformers
# torch
